Use an a[role="presentation"] instead of div and onClick
Using a div with onClick prevents middle click (to open in a new tab)
to no longer work correctly, because the browser doesn't understand
that it's a link.

We want to keep "Read Guide" as the only focusable/semantic link
to open the guide, but mouse users should still be able to click on
the entire card (except for the subclass icon) to open the guide
link.